requirements.txt
opencv-python==4.11.0.86
numpy==2.2.4
google-generativeai>=0.3.1
pillow>=10.0.1
streamlit>=1.28.0 